+++ app/Helper.php
@@ -7,12 +7,15 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Iatstuti\Database\Support\NullableFields;
 use OwenIt\Auditing\Contracts\Auditable;
 use Carbon\Carbon;
+use AustinHeap\Database\Encryption\Traits\HasEncryptedAttributes;
 
 class Helper extends Model implements Auditable
 {
     use SoftDeletes;
     use NullableFields;
     use \OwenIt\Auditing\Auditable;
+    use HasEncryptedAttributes;
+
 
     /**
      * Get the person record associated with the helper.
@@ -85,6 +88,7 @@ class Helper extends Model implements Auditable
         'shoe_size',
         'urgent_needs',
         'work_needs',
+        'notes',
     ];
 
     /**
@@ -174,4 +178,16 @@ class Helper extends Model implements Auditable
         return $this->is_active && $this->work_trial_period;
     }
 
+    /**
+     * The attributes that should be encrypted on save.
+     *
+     * @var array
+     */
+    protected $encrypted = [
+        'casework_vulnerability',
+        'casework_lawyer_name',
+        'casework_lawyer_phone',
+        'casework_lawyer_email',
+        'notes',
+    ];
 }